Moved and renamed main ui file from ui_main to ui, in the ui/ folder. Modified __main__ to call the MainApplication class upon exec.
