+++ project/__main__.py
@@ -1,0 +1,13 @@
+import sys
+from PySide2.QtWidgets import QApplication
+from ui.ui import MainApplication
+
+
+def main():
+    app = QApplication(sys.argv)
+    window = MainApplication()
+    sys.exit(app.exec_())
+
+
+if __name__ == "__main__":
+    main()
